// frontend-admin/src/api/generated/model/adminOfflineOrderRowDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export interface AdminOfflineOrderRowDto {
  cashRegisterId?: string;
  /** @nullable */
  cashRegisterLabel?: string | null;
  createdAtUtc?: string;
  /** @nullable */
  errorMessage?: string | null;
  hoursRemaining?: number;
  id?: string;
  /** @nullable */
  offlineOrderId?: string | null;
  orderTotal?: number;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  status?: string | null;
  syncAttempts?: number;
  /** @nullable */
  syncedInvoiceNumber?: string | null;
  /** @nullable */
  syncedPaymentId?: string | null;
}