// frontend-admin/src/api/generated/model/getApiAdminOfflineOrdersParams.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API â€” RKSV-compliant POS / Admin / shared Auth surfaces
 * OpenAPI spec version: v1
 */

export type GetApiAdminOfflineOrdersParams = {
status?: string;
cashRegisterId?: string;
pageNumber?: number;
pageSize?: number;
};